# internal/dto/user_dto.go
package dto

type UserUpdateInfoReqDto struct {
	Nickname  *string `json:"nickname" form:"nickname" binding:"omitempty,min=2,max=20"`
	AvatarURL *string `json:"avatarURL" form:"avatarURL" binding:"omitempty,url"`
	Desc      *string `json:"desc" form:"desc" binding:"omitempty,max=200"`
	Active    *bool   `json:"active" form:"active"`
}

type UserInfoResDto struct {
	UniCode     string   `json:"uniCode"`
	Active      bool     `json:"active"`
	Email       string   `json:"email"`
	Mobile      string   `json:"mobile"`
	CountryCode string   `json:"countryCode"`
	Desc        string   `json:"desc"`
	Nickname    string   `json:"nickname"`
	AvatarURL   string   `json:"avatarURL"`
	Roles       []string `json:"roles"`
}

type UserTableQueryReqDto struct {
	PaginationReqDto
	Nickname    *string `json:"nickname" form:"nickname"`
	Email       *string `json:"email" form:"email"`
	Mobile      *string `json:"mobile" form:"mobile"`
	CountryCode *string `json:"countryCode" form:"countryCode"`
}

type UserListItemResDto struct {
	ID          uint64   `json:"id"`
	CreatedAt   string   `json:"createdAt"`
	UniCode     string   `json:"uniCode"`
	Active      bool     `json:"active"`
	Email       string   `json:"email"`
	Mobile      string   `json:"mobile"`
	CountryCode string   `json:"countryCode"`
	Desc        string   `json:"desc"`
	Nickname    string   `json:"nickname"`
	AvatarURL   string   `json:"avatarURL"`
	Roles       []string `json:"roles"`
}

// CreateUserReqDto is the request DTO for admin user creation.
type CreateUserReqDto struct {
	Email    string `json:"email" binding:"required,email"`
	Nickname string `json:"nickname" binding:"required,min=2,max=20"`
}

type UserSetRolesReqDto struct {
	RoleIDs []uint64 `json:"roleIds"`
}